# lecture/go-swagger/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/acc/v01/ok": {
            "get": {
                "description": "api test를 위한 기능.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "call GetOK, return ok by json.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User name",
                        "name": "name",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controller.Controller"
                        }
                    }
                }
            }
        },
        "/acc/v01/person/add": {
            "post": {
                "description": "Person 데이터를 추가하는 기능",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "call CtlAddPerson body \"name\", \"age\", \"pnum\", return object by json.",
                "parameters": [
                    {
                        "description": "person",
                        "name": "person",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.Person"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Person"
                        }
                    }
                }
            }
        },
        "/acc/v01/person/update": {
            "put": {
                "description": "Person 데이터를 업데이트하는 기능",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "call CtlUpdatePerson body \"name\", \"age\" return objectId",
                "parameters": [
                    {
                        "description": "data",
                        "name": "name",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.Person"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/acc/v01/person/{name}": {
            "get": {
                "description": "이름으로 원하는 person의 데이터를 찾는 기능",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "call CtlGetByName with path parameter \"name\", return object by json.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "name",
                        "name": "name",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Person"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "controller.Controller": {
            "type": "object"
        },
        "model.Person": {
            "type": "object",
            "properties": {
                "age": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "pnum": {
                    "type": "string"
                }
            }
        }
    }
}